pull Collection type to common/types
