+++ monolith/interactors/getFormatLegalities.ts
@@ -1,8 +1,6 @@
-import collectionFromLocation, {
-    Collection,
-} from '../lib/collectionFromLocation';
 import getDatabaseConnection from '../database';
-import { ClubhouseLocation } from '../common/types';
+import { ClubhouseLocation, Collection } from '../common/types';
+import collectionFromLocation from '../lib/collectionFromLocation';
 
 async function getFormatLegalities(location: ClubhouseLocation) {
     try {
