+++ monolith/interactors/getReceivingById.ts
@@ -1,9 +1,7 @@
 import { ObjectID } from 'mongodb';
-import { ClubhouseLocation } from '../common/types';
+import { ClubhouseLocation, Collection } from '../common/types';
 import getDatabaseConnection from '../database';
-import collectionFromLocation, {
-    Collection,
-} from '../lib/collectionFromLocation';
+import collectionFromLocation from '../lib/collectionFromLocation';
 
 async function getReceivingById(id: string, location: ClubhouseLocation) {
     try {
